Initial public version of community lock core
I created the original design late in 2020 but this is the first version
that I'm going to publish.

The Blender file includes a script to export the Community Core design as
STL files. The exported files will be written to subdirectory "stl-export"
and any existing files in that directory will be overwritten without a
warning.

The initial model has all the disks and spacers as a single object to make
it easier to apply modifiers and change the disks and spacers. This is
reflected in STL export where all the disks are in one file and all the
spacers are in another file.