Add another design version, v21
This design mostly adds just configurable keycuts. To configure the keycuts
adjust the custom properties of object ".KEYCUT".

The quality of curves is now easily configurable, too. To change it, adjust
custom properties of object ".QUALITY".

After setting the required quality, you can switch to "Scripting" tab
on the the top and click the "Play" button ("Run Script") or press Alt+P
to export all the parts into STL files. I think all the parts are now
fully working and manifold but I don't have the equipment to manufacture
any of the parts.