Add design v27
This variant of the design avoids using plug/cylinder and tries to create
working DD core with just disks, spacers and 2 bars.

I think this design is vulnerable to manufacturing tolerances because
attacker could push the code disk towards the sidebar and rotate the
sidebar and the manufacturing tolerances couldn't prevent offsetting
the code disk away from the true center of the lock housing and as a
result, the attacker could feel the side of the disk against the sidebar.

I think this could be a good design with insanely tight tolerances but
that's not something I'm trying to create.