Add design v29
This variant of the design avoids using plug/cylinder and tries to create
working DD core with just disks, spacers and single sidebar.

With this design I've tried to consider how hard it would be manufacture
all the parts. The part that I cannot figure out is how to easily
build the labyrinth part of the front face where sidebar hole (if simply
drilled) would intersect with the hole needed for the keyway at the
front of the lock. This would result a section of lock without labyrinth
section to access the sidebar from keyway. I'm not sure if this is a
great vulnerability because accessing the actual code disks this way
would require inserting the tool between the front face of the lock and
the first spacer and then feeling the code disks next to sidebar.
However, I'd prefer coming up with a design that doesn't leave such
direct access to sidebar.

Also, this design uses only 1 mm thick sidebar which would be bent. As the
available space is only 0.8 mm I'm hoping that the sidebar cannot be
compressed 0.2 mm which would allow brute forcing the lock open.

I think this design would need one extra disk at the front to implement
the labyrinth surface and hold the sidebar straight similar to rear
spacer near the end plug.

I'm also pretty sure that the circlip wouldn't fit into the position
in real world because one needs to be able to press the circlip smaller
to fit it into the groove and there's probably not enough space to
actually put the circlip on place.